+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/ribbon/PolarisServerList.java
@@ -56,7 +56,8 @@ public class PolarisServerList extends AbstractServerList<Server> {
 	}
 
 	private List<Server> getServers() {
-		InstancesResponse allInstances = polarisDiscoveryHandler.getFilteredInstances(serviceId);
+		InstancesResponse allInstances = polarisDiscoveryHandler
+				.getFilteredInstances(serviceId);
 		ServiceInstances serviceInstances = allInstances.toServiceInstances();
 		List<Server> polarisServers = new ArrayList<>();
 		for (Instance instance : serviceInstances.getInstances()) {