+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/registry/PolarisServiceRegistryAutoConfigurationTest.java
@@ -17,7 +17,7 @@
 
 package com.tencent.cloud.polaris.registry;
 
-import com.tencent.cloud.polaris.context.PolarisContextConfiguration;
+import com.tencent.cloud.polaris.context.PolarisContextAutoConfiguration;
 import com.tencent.cloud.polaris.discovery.PolarisDiscoveryAutoConfiguration;
 import com.tencent.cloud.polaris.discovery.PolarisDiscoveryClientConfiguration;
 import com.tencent.polaris.test.mock.discovery.NamingServer;
@@ -46,9 +46,10 @@ public class PolarisServiceRegistryAutoConfigurationTest {
 	private static NamingServer namingServer;
 
 	private WebApplicationContextRunner contextRunner = new WebApplicationContextRunner()
-			.withConfiguration(AutoConfigurations.of(PolarisContextConfiguration.class,
-					PolarisServiceRegistryAutoConfiguration.class,
-					PolarisDiscoveryClientConfiguration.class))
+			.withConfiguration(
+					AutoConfigurations.of(PolarisContextAutoConfiguration.class,
+							PolarisServiceRegistryAutoConfiguration.class,
+							PolarisDiscoveryClientConfiguration.class))
 			.withPropertyValues("spring.application.name=" + SERVICE_PROVIDER)
 			.withPropertyValues("server.port=" + PORT)
 			.withPropertyValues("spring.cloud.polaris.address=grpc://127.0.0.1:10081");