+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/PolarisContextProperties.java
@@ -55,6 +55,11 @@ public class PolarisContextProperties {
 	private String localIpAddress;
 
 	/**
+	 * If polaris enabled.
+	 */
+	private Boolean enabled;
+
+	/**
 	 * polaris namespace.
 	 */
 	private String namespace = "default";
@@ -111,6 +116,14 @@ public class PolarisContextProperties {
 		this.localIpAddress = localIpAddress;
 	}
 
+	public Boolean getEnabled() {
+		return enabled;
+	}
+
+	public void setEnabled(Boolean enabled) {
+		this.enabled = enabled;
+	}
+
 	public String getNamespace() {
 		return namespace;
 	}
@@ -126,4 +139,5 @@ public class PolarisContextProperties {
 	void setService(String service) {
 		this.service = service;
 	}
+
 }