+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - [Bugfix: fix circuitbreaker http code greater than 400 as fail response bug](https://github.com/Tencent/spring-cloud-tencent/pull/117)
 - [Feat: optimize router dependency](https://github.com/Tencent/spring-cloud-tencent/pull/114)
 - [Refactor: refactor transfer metadata](https://github.com/Tencent/spring-cloud-tencent/pull/112)
+- [feat:add switch of polaris, discovery and register.](https://github.com/Tencent/spring-cloud-tencent/pull/133)